chore: refactoring, move vesting genesis to a new pkg
